Refactor: Worker module
