+++ internal/origins/drivers/brew/fetch_casks.go
@@ -5,8 +5,8 @@ import (
 	"os"
 	"path/filepath"
 	"qp/internal/origins/shared"
-	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
+	"qp/internal/worker"
 	"sync"
 )
