+++ internal/origins/drivers/deb/parser.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"qp/internal/consts"
 	"qp/internal/origins/shared/debstyle"
-	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
+	"qp/internal/worker"
 	"strconv"
 	"strings"
 	"sync"
