+++ internal/origins/drivers/flatpak/discover.go
@@ -3,8 +3,8 @@ package flatpak
 import (
 	"os"
 	"path/filepath"
-	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
+	"qp/internal/worker"
 	"strings"
 	"sync"
 )
