+++ internal/origins/drivers/opkg/parser.go
@@ -8,8 +8,8 @@ import (
 	"path/filepath"
 	"qp/internal/consts"
 	"qp/internal/origins/shared/debstyle"
-	"qp/internal/origins/worker"
 	"qp/internal/pkgdata"
+	"qp/internal/worker"
 	"strconv"
 	"strings"
 	"sync"
